Added tests for UserDetailsViewModel Cleanup User models Fixed broken tests

+++ network/src/main/java/com/slide/test/network/model/PostDto.kt
@@ -6,9 +6,9 @@ import com.squareup.moshi.Json
  * Created by Stefan Halus on 11 Jun 2025
  */
 data class PostDto(
-    val id: Int,
+    val id: Long,
     @Json(name = "user_id")
-    val userId: Int,
+    val userId: Long,
     val title: String,
     val body: String
 )
